backend/requirements.txt
alembic==1.5.8
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-futures==1.0.0
six==1.15.0
SQLAlchemy==1.4.3
tqdm==4.59.0
urllib3==1.26.4
Werkzeug==1.0.1
yahooquery==2.2.15
